% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_fake.R
\name{generate_fake_compagnies}
\alias{generate_fake_compagnies}
\title{Title}
\usage{
generate_fake_compagnies(n_rows)
}
\arguments{
\item{n_rows}{(int) number of rows to get}
}
\value{
a tibble
}
\description{
Title
}
